My mind was wondering around a lot on this commit. It'll be a bit tricky to do all the code changes justice.
The primary idea with these changes is to make it possible to create the project
and then the environment, including installing mongrel2 and zeromq into a custom
virtualenv.

After installing BPM you could use it like this:

    $ pip install bpm
    $ bpm project create
    $ cd brubeck_project
    $ bpm env create

The environment is now complete. But to enable it, you still need one more
command:

    $ source bin/bpmenv

That takes care of setting up everything required for the brubeck project
environment. It also allows a user to extend the environment easily, since
that file is in the project and only copied from bpm.

Documentation still to come.