This commit modifies the project creation process. The primary goal of the modification is to sweep across a newly create project and render template variables.
At the time of this commit, the only variable used is
`BPM_PROJECT_NAME` which gets replaced with the name of the Brubeck
project being created.

The template system lives in `bpm.system` and offers three functions.

1. Find files that have template tags in them
2. Render an individual file
3. Render a whole directory

The effects of this patch can be observed when using the command below

    $ bpm project create [-n <project name>]